+++ src/module.d.ts
@@ -547,6 +547,7 @@ declare namespace posthog {
         inapp_protocol?: string
         inapp_link_new_window?: boolean
         request_batching?: boolean
+        sanitize_properties?: (properties: Record<string, any>, event_name: string) => Record<string, any>
     }
 
     interface OptInOutCapturingOptions {
@@ -680,6 +681,8 @@ declare namespace posthog {
 
     export class feature_flags {
         static getFlags(): string[]
+
+        static reloadFeatureFlags(callback: (flags: string[]) => any): void
 
         /*
          * See if feature flag is enabled for user.